Icons added to the user Profile section of profiles. UserCard was moved to containers folder. Some unused code was removed, and components added as imports.
