+++ client/src/components/CurrentUserCard/index.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Card, Icon, Image } from 'semantic-ui-react';
+import { Card, Icon, Image, Container, Header } from 'semantic-ui-react';
 
 import _ from 'lodash';
 
@@ -7,6 +7,9 @@ class CurrentUserContainer extends Component {
 
   render() {
     return (
+      <Container>
+        <br />
+        <Header as='h3' icon><Icon name='user circle' />User Profile</Header>
         <Card color='olive' >
           <Image src={this.props.currentUser.badge} wrapped ui={false} />
           <Card.Content>
@@ -20,41 +23,15 @@ class CurrentUserContainer extends Component {
               {this.props.currentUser.bio}
             </Card.Description>
           </Card.Content>
-          <Card.Content  extra>
-              <Icon name='user' />
-              {this.props.currentUser.email}
+          <Card.Content extra>
+            <Icon name='user' />
+            {this.props.currentUser.email}
           </Card.Content>
         </Card>
-      
-
+      </Container>
     )
   }
 };
 
-
 export default CurrentUserContainer;
 
-
-
-
-
-
-// import React from 'react';
-// import Match from '../CurrentUser';
-
-// function MatchContainer({ id, badge, first_name, last_name, strength, weakness, bio, email }) {
-//     return(
-//         <div>
-//             <Match id={id}
-//             badge={badge}
-//             firstName={first_name}
-//             lastName={last_name}
-//             strength={strength}
-//             weakness={weakness}
-//             bio={bio}
-//             email={email} />
-//         </div>
-//     )
-// }
-
-// export default MatchContainer;
