+++ client/src/components/Loader/index.js
@@ -6,7 +6,6 @@ const Loader = () => (
     <Dimmer active>
       <Loader />
     </Dimmer>
-
     <Image src='https://react.semantic-ui.com/images/wireframe/short-paragraph.png' />
   </Segment>
 )
